requirements.txt
python-dotenv==1.1.0
requests==2.32.4
uvicorn==0.35.0
fastapi==0.116.1
pyyaml==6.0.2
rich==14.1.0